+++ dice_maiden_init.rb
@@ -1,7 +1,7 @@
 #!/usr/bin/ruby
 # simple init script for dice maiden shards
 # Due to how long discordrb takes to start, starting multiple shards at once can cause them to hang
-# This script spaces the shard inits out every minute to help prevent this
+# This script spaces the shard inits out to help prevent this
 
 i = 0
 
@@ -19,5 +19,5 @@ loop do
   `bluepill dice_maiden start dice_maiden#{i}`
   i += 1
   break if i == ARGV[0].to_i
-  sleep 30
+  sleep 10
 end